Move mesh function definitions into IMesh.cpp
#741
